fix polytomouskey delete, if root node is null

+++ cdmlib-services/src/main/java/eu/etaxonomy/cdm/api/service/PolytomousKeyServiceImpl.java
@@ -164,7 +164,9 @@ public class PolytomousKeyServiceImpl extends IdentifiableServiceBase<Polytomous
 
 	    if (result.isOk()){
     	    try{
-    	        nodeService.delete(root.getUuid(), true);
+    	        if (root != null){
+    	            nodeService.delete(root.getUuid(), true);
+    	        }
     	    }catch(Exception e){
     	        result.addException(e);
     	        result.setAbort();
